updated simple example
