+++ examples/simple/README.md
@@ -1,0 +1,19 @@
+# Simple Example
+
+In this example, we simply put and get a value from HollowDB.
+
+```sh
+yarn install
+```
+
+Run the example:
+
+```sh
+node index.js
+```
+
+You can also use a bundler (Irys) with, if you have funded your wallet:
+
+```sh
+USE_IRYS=true node index.js
+```
