+++ examples/simple/package.json
@@ -3,8 +3,8 @@
   "author": "FirstBatch Team <dev@firstbatch.xyz>",
   "private": true,
   "dependencies": {
-    "@bundlr-network/client": "^0.11.0",
-    "hollowdb": "latest",
+    "@irys/sdk": "^0.1.1",
+    "hollowdb": "^1.3.5",
     "hollowdb-prover": "latest"
   }
 }
